+++ packages/saas-ui-data-table/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @saas-ui/data-table
 
+## 1.0.0-rc.7
+
+### Patch Changes
+
+- 1744543: Updated dependencies.
+- Updated dependencies [1744543]
+  - @saas-ui/layout@1.0.0-rc.7
+
 ## 1.0.0-rc.6
 
 ### Patch Changes