+++ packages/saas-ui-hotkeys/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @saas-ui/hotkeys
 
+## 1.0.0-rc.4
+
+### Patch Changes
+
+- 1744543: Updated dependencies.
+- Updated dependencies [1744543]
+  - @saas-ui/search-input@1.0.0-rc.4
+
 ## 1.0.0-rc.3
 
 ### Patch Changes