+++ packages/saas-ui-password-input/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @saas-ui/password-input
 
+## 1.0.0-rc.5
+
+### Patch Changes
+
+- 1744543: Updated dependencies.
+- Updated dependencies [1744543]
+  - @saas-ui/input-right-button@1.0.0-rc.5
+
 ## 1.0.0-rc.4
 
 ### Patch Changes